Add files via upload
Here we are. The first scripts of myself. I wrote them with Gemini Advanced for my purposes. These are the local network ip scanners which returns all the results with an output.txt. It's even includes MAC Adresses and , if exist, their vendors. If you have connection you can use the online one with itself. But if you do not have WAN connection you can use the offline one with the mac-vendor.txt (which taken from :https://gist.github.com/aallan) for vendors. Bored of installing the shitty and clunky softwares to scan networks and even sometimes systems can delete them claiming theyre viruses. LOL. So, we needed a script for it for our infastructure and our clients. Here we are. Enjoy.